Avoid running glGetString during import, but it's too fragile
Basically anything which takes an extension or similar function and 
does a boolean test on it is going to cause a glGetString() operation,
and that will crash on OS-X. There is definitely some code in the 
images module that does bool(imageFunction) where it really should
not be doing so (i.e. on import).  That should work now, simply because
the functions *happen* to be core (version XXX).