Remove duplicate wrapping for glGetBufferParameteriv, guard against new errors on scalar unpack
The changes back in 2015 that made it possible to have N wrappers do not
guard against the *same* parameter being wrapped twice. The result in that
case is to crash with a rather opaque error deep within the wrapper sets.

Also guards against an apparent change in numpy where we no longer can use
shape (1,) to mean that a scalar can be unpacked? The result of that is that
we get IndexError instead of TypeError when trying to unpack the single-value
scalar for oldStyleReturn.