FIX testsuite: Further fixes for testsuite to run in extremely old OpenGL
Running Windows VMS with their 1.x OpenGL implementations would cause
lots of failures in the test suite. This patches the suite, but it
does it with ugly code that just exits the test when the feature isn't
available. Should replace with run-time (can't be load-time) checks
for features with reasonable reporting via pytest skip