GH#38 WIP Renaming of parameters by upstream in multiple APIs (current registry)
This updates our wrappers to use the current khronos registry mainline,
but it reflects the parameter-name changes, so there my be regressions
that need to be tracked down with custom wrappers.

User code accessing the APIs with changed parameters might be affected
IFF they have used keyword-parameters for the changed parameter names.