GH #47 Do not incref numpy arrays on being handed into the GL
Note, this is almost certainly going to result in some client
code failing, as the previous code was hacking around the issues
where python code would create an array, pass it into a function
that required a pointer that lives as long as the binding,
and then garbage collect that pointer.  But that hack was wrong,
so let's be correct.