also check for .so without version extension
bookworm and later do not ship the freeglut3 library any more that provided the versioned symlink libglut.so.3 for the libglut.so.3.12.0 library, but there is libglut.so , still, and libglut.so.3.12.

$ ls -l /usr/lib/x86_64-linux-gnu/*glut*
-rw-r--r-- 1 root root 654216  5. Nov 21:59 /usr/lib/x86_64-linux-gnu/libglut.a
lrwxrwxrwx 1 root root     15  5. Nov 21:59 /usr/lib/x86_64-linux-gnu/libglut.so -> libglut.so.3.12
lrwxrwxrwx 1 root root     17  5. Nov 21:59 /usr/lib/x86_64-linux-gnu/libglut.so.3.12 -> libglut.so.3.12.0
-rw-r--r-- 1 root root 356016  5. Nov 21:59 /usr/lib/x86_64-linux-gnu/libglut.so.3.12.0